Minor improvements to fix Qulice errors
