+++ src/main/java/org/eolang/jetbrains/EoExternalAnnotator.java
@@ -1,24 +1,44 @@
 /*
  * The MIT License (MIT)
  *
+ *
+ *
  * Copyright (c) 2021-2022 Stepan Strunkov
  *
+ *
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
+ *
  * of this software and associated documentation files (the "Software"), to deal
+ *
  * in the Software without restriction, including without limitation the rights
+ *
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *
  * copies of the Software, and to permit persons to whom the Software is
+ *
  * furnished to do so, subject to the following conditions:
  *
+ *
+ *
  * The above copyright notice and this permission notice shall be included
+ *
  * in all copies or substantial portions of the Software.
  *
+ *
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *
  * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ *
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *
  * SOFTWARE.
  */
 
@@ -31,7 +51,6 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import java.util.ArrayList;
 import java.util.List;
-
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 
@@ -40,12 +59,9 @@ import org.jetbrains.annotations.NotNull;
  * elements with errors or warnings. Because such analysis can be expensive, we don't want it in the
  * GUI event loop. Jetbrains provides this external annotator mechanism to run these analyzers out
  * of band.
- * @since 0.0.0
- * @checkstyle MultilineJavadocTagsCheck (200 lines)
- * @checkstyle MemberNameCheck (200 lines)
  */
 public class EoExternalAnnotator
-        extends ExternalAnnotator<PsiFile, List<EoExternalAnnotator.Issue>> {
+    extends ExternalAnnotator<PsiFile, List<EoExternalAnnotator.Issue>> {
 
     // Called first; in our case, just return file and do nothing.
     @Override
@@ -69,15 +85,16 @@ public class EoExternalAnnotator
 
     // Called 3rd to actually annotate the editor window.
     @Override
+    @SuppressWarnings("PMD:SystemPrintln")
     public final void apply(@NotNull final PsiFile file, final @NotNull List<Issue> issues,
         @NotNull final AnnotationHolder holder) {
         for (final Issue issue : issues) {
-            final TextRange range = issue.getOffendingNode().getTextRange();
+            final TextRange range = issue.getOffendnode().getTextRange();
             if (range.getStartOffset() > range.getEndOffset()) {
-                System.err.println("Invalid TextRange: " + range);
+                System.err.println("Invalid TextRange");
                 continue;
             }
-                holder.createErrorAnnotation(range, issue.getMsg());
+            holder.createErrorAnnotation(range, issue.getMsg());
         }
     }
 
@@ -95,7 +112,7 @@ public class EoExternalAnnotator
         /**
          * Node.
          */
-        private final PsiElement offendingNode;
+        private final PsiElement offendnode;
 
         /**
          * Issue init.
@@ -104,7 +121,7 @@ public class EoExternalAnnotator
          */
         public Issue(final String msg, final PsiElement node) {
             this.msg = msg;
-            this.offendingNode = node;
+            this.offendnode = node;
         }
 
         /**
@@ -119,8 +136,8 @@ public class EoExternalAnnotator
          * Accessor.
          * @return Offending node
          */
-        final PsiElement getOffendingNode() {
-            return this.offendingNode;
+        final PsiElement getOffendnode() {
+            return this.offendnode;
         }
     }
 }
