+++ src/main/java/org/eolang/jetbrains/EoSyntaxHighlighter.java
@@ -1,24 +1,44 @@
 /*
  * The MIT License (MIT)
  *
+ *
+ *
  * Copyright (c) 2021-2022 Stepan Strunkov
  *
+ *
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
+ *
  * of this software and associated documentation files (the "Software"), to deal
+ *
  * in the Software without restriction, including without limitation the rights
+ *
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *
  * copies of the Software, and to permit persons to whom the Software is
+ *
  * furnished to do so, subject to the following conditions:
  *
+ *
+ *
  * The above copyright notice and this permission notice shall be included
+ *
  * in all copies or substantial portions of the Software.
  *
+ *
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *
  * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ *
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *
  * SOFTWARE.
  */
 // @checkstyle AvoidStaticImportCheck (15 lines)
@@ -149,14 +169,9 @@ public class EoSyntaxHighlighter extends SyntaxHighlighterBase {
             case EoLexer.RSQ:
             case EoLexer.SLASH:
             case EoLexer.CONST:
-            //case EoLexer.DOTS:
             case EoLexer.COLON:
-            //case EoLexer.COPY:
             case EoLexer.ARROW:
-            //case EoLexer.VERTEX:
-            //case EoLexer.SIGMA:
             case EoLexer.XI:
-            //case EoLexer.AT:
             case EoLexer.RHO:
             case EoLexer.PLUS:
             case EoLexer.MINUS:
@@ -164,7 +179,6 @@ public class EoSyntaxHighlighter extends SyntaxHighlighterBase {
             case EoLexer.STAR:
                 key = EoSyntaxHighlighter.KEYWORD;
                 break;
-            //case EoLexer.BOOL:
             case EoLexer.INT:
             case EoLexer.FLOAT:
             case EoLexer.HEX:
@@ -185,10 +199,6 @@ public class EoSyntaxHighlighter extends SyntaxHighlighterBase {
             case EoLexer.DOT:
                 key = EoSyntaxHighlighter.DOT;
                 break;
-            /*case EoLexer.COMMENT:
-                key = EoSyntaxHighlighter.COMMENT;
-              break;
-            */
             case EoLexer.HASH:
                 key = EoSyntaxHighlighter.HASH;
                 break;
