--- src/test/java/org/eolang/jetbrains/EoColorSettingsPageTest.java
@@ -1,36 +1,0 @@
-package org.eolang.jetbrains;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class EoColorSettingsPageTest {
-
-    @Test
-    void getAdditionalHighlightingTagToDescriptorMap() {
-    }
-
-    @Test
-    void getIcon() {
-    }
-
-    @Test
-    void getHighlighter() {
-    }
-
-    @Test
-    void getDemoText() {
-    }
-
-    @Test
-    void getAttributeDescriptors() {
-    }
-
-    @Test
-    void getColorDescriptors() {
-    }
-
-    @Test
-    void getDisplayName() {
-    }
-}
