--- src/test/java/org/eolang/jetbrains/EoCompileActionTest.java
@@ -1,12 +1,0 @@
-package org.eolang.jetbrains;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class EoCompileActionTest {
-
-    @Test
-    void actionPerformed() {
-    }
-}
